Keys output operator in log.hpp/cpp
