+++ src/Keys.cpp
@@ -1,5 +1,4 @@
 #include "Keys.hpp"
-#include <iomanip>
 
 Keys::Keys(uint32 x, uint32 y, uint32 z)
  : x(x), y(y), z(z)
@@ -24,13 +23,4 @@ void Keys::updateBackward(const bytevec& ciphertext, std::size_t current, std::s
 
     for(rit i = rit(ciphertext.begin() + current); i != rit(ciphertext.begin() + target); ++i)
         updateBackward(*i);
-}
-
-std::ostream& operator<<(std::ostream& os, const Keys& keys)
-{
-    return os << std::hex
-              << std::setw(8) << keys.getX() << " "
-              << std::setw(8) << keys.getY() << " "
-              << std::setw(8) << keys.getZ()
-              << std::dec;
 }
